moved identifier convertion functions to ontologies.standardization.py
