+++ src/napistu/constants.py
@@ -494,98 +494,3 @@ SOURCE_STANDARD_COLUMNS = EXPECTED_PW_INDEX_COLUMNS | {
     SOURCE_SPEC.ENTRY,
     SOURCE_SPEC.MODEL,
 }
-
-# rules for specific ontologies
-ONTOLOGIES = SimpleNamespace(
-    BIGG_METABOLITE="bigg_metabolite",
-    BIORXIV="biorxiv",
-    CHEBI="chebi",
-    CHEMSPIDER="chemspider",
-    CORUM="corum",  # protein complexes
-    DRUGBANK="drugbank",
-    DOI="doi",  # doi.org digital object identifiers
-    DX_DOI="dx_doi",  # dx.doi.org digital object identifiers
-    EC_CODE="ec-code",  # enzyme commission codes
-    EBI_REFSEQ="ebi_refseq",  # ebi DNA reference sequences
-    ENSEMBL_GENE="ensembl_gene",
-    ENSEMBL_GENE_VERSION="ensembl_gene_version",
-    ENSEMBL_TRANSCRIPT="ensembl_transcript",
-    ENSEMBL_TRANSCRIPT_VERSION="ensembl_transcript_version",
-    ENSEMBL_PROTEIN="ensembl_protein",
-    ENSEMBL_PROTEIN_VERSION="ensembl_protein_version",
-    ENVIPATH="envipath",  # microbial environmental transformation pathways
-    GENE_NAME="gene_name",
-    GENOME_NET="genome_net",
-    GO="go",  # gene ontology
-    GUIDETOPHARMACOLOGY="guidetopharmacology",
-    INCHIKEY="inchikey",  # hashed chemical structure
-    INTACT="intact",  # intact protein interactions
-    KEGG="kegg",
-    KEGG_DRUG="kegg.drug",
-    MIRBASE="mirbase",  # microRNAs
-    MATRIXDB_BIOMOLECULE="matrixdb_biomolecule",  # extracellular matrix molecules and interactions
-    MATRIXDB_MOLECULE_CLASS="matrixdb_molecule_class",  # extracellular matrix molecules and interactions
-    MDPI="mdpi",  # mdpi journal articles
-    NCBI_BOOKS="ncbi_books",
-    NCBI_ENTREZ_GENE="ncbi_entrez_gene",
-    NCBI_ENTREZ_PCCOMPOUND="ncbi_entrez_pccompound",  # pubchem inchikeys
-    NCBI_REFSEQ="ncbi_refseq",  # ncbi reference sequences
-    NCI_THESAURUS="nci_thesaurus",
-    OLS="ols",  # ontology of ontologies
-    PHAROS="pharos",  # pharos gene summaries
-    PHOSPHOSITE="phosphosite",  # phosphosite.org kinase and ligand interactions
-    PUBCHEM="pubchem",
-    PUBMED="pubmed",
-    REACTOME="reactome",
-    RHEA="rhea",  # curated metabolic reactions
-    RNACENTRAL="rnacentral",
-    SGC="sgc",  # structural genomics consortium
-    SGD="sgd",  # saccharomyces genome database
-    SIGNOR="signor",  # signaling pathways
-    SMILES="smiles",  # molecular structure
-    SYMBOL="symbol",
-    URL="url",
-    UNIPROT="uniprot",
-    WIKIPATHWAYS="wikipathways",
-)
-
-ONTOLOGIES_LIST = list(ONTOLOGIES.__dict__.values())
-
-ONTOLOGY_SPECIES_ALIASES = {
-    ONTOLOGIES.NCBI_ENTREZ_GENE: {"ncbigene", "ncbi_gene"},
-    ONTOLOGIES.ENSEMBL_GENE: {"ensembl_gene_id"},
-    ONTOLOGIES.UNIPROT: {"Uniprot"},
-    ONTOLOGIES.CORUM: {"CORUM"},
-    ONTOLOGIES.SIGNOR: {"SIGNOR"},
-}
-
-ONTOLOGY_PRIORITIES = pd.DataFrame(
-    [
-        {IDENTIFIERS.ONTOLOGY: ONTOLOGIES.REACTOME, "ontology_rank": 1},
-        {IDENTIFIERS.ONTOLOGY: ONTOLOGIES.ENSEMBL_GENE, "ontology_rank": 2},
-        {IDENTIFIERS.ONTOLOGY: ONTOLOGIES.CHEBI, "ontology_rank": 3},
-        {IDENTIFIERS.ONTOLOGY: ONTOLOGIES.UNIPROT, "ontology_rank": 4},
-        {IDENTIFIERS.ONTOLOGY: ONTOLOGIES.GO, "ontology_rank": 5},
-    ]
-)
-
-ENSEMBL_MOLECULE_TYPES_TO_ONTOLOGY = {
-    "G": ONTOLOGIES.ENSEMBL_GENE,
-    "T": ONTOLOGIES.ENSEMBL_TRANSCRIPT,
-    "P": ONTOLOGIES.ENSEMBL_PROTEIN,
-}
-
-ENSEMBL_MOLECULE_TYPES_FROM_ONTOLOGY = {
-    ONTOLOGIES.ENSEMBL_GENE: "G",
-    ONTOLOGIES.ENSEMBL_TRANSCRIPT: "T",
-    ONTOLOGIES.ENSEMBL_PROTEIN: "P",
-}
-
-ENSEMBL_SPECIES_FROM_CODE = {"MUS": "Mus musculus"}
-ENSEMBL_SPECIES_TO_CODE = {"Mus musculus": "MUS"}
-
-ENSEMBL_PREFIX_TO_ONTOLOGY = {
-    "ENSG": ONTOLOGIES.ENSEMBL_GENE,
-    "ENST": ONTOLOGIES.ENSEMBL_TRANSCRIPT,
-    "ENSP": ONTOLOGIES.ENSEMBL_PROTEIN,
-}
