+++ src/napistu/matching/mount.py
@@ -5,7 +5,7 @@ from typing import Dict, Optional, Set, Union
 import pandas as pd
 
 from napistu import identifiers, utils
-from napistu.constants import ONTOLOGIES_LIST, SBML_DFS
+from napistu.constants import SBML_DFS
 from napistu.matching.constants import (
     BIND_DICT_OF_WIDE_RESULTS_STRATEGIES,
     BIND_DICT_OF_WIDE_RESULTS_STRATEGIES_LIST,
@@ -14,6 +14,7 @@ from napistu.matching.constants import (
     RESOLVE_MATCHES_TMP_WEIGHT_COL,
 )
 from napistu.matching.species import match_features_to_wide_pathway_species
+from napistu.ontologies.constants import ONTOLOGIES_LIST
 
 # Type annotations use string literals to avoid circular imports
