+++ src/tests/test_consensus.py
@@ -6,7 +6,18 @@ import os
 import pandas as pd
 import pytest
 
-from napistu import consensus, indices, sbml_dfs_core, source
+from napistu.consensus import (
+    _build_consensus_identifiers,
+    _create_consensus_sources,
+    _pre_consensus_compartment_check,
+    _pre_consensus_ontology_check,
+    _reduce_to_consensus_ids,
+    _report_consensus_merges,
+    _update_foreign_keys,
+    construct_consensus_model,
+    construct_sbml_dfs_dict,
+    prepare_consensus_model,
+)
 from napistu.constants import (
     BQB,
     IDENTIFIERS,
@@ -17,8 +28,11 @@ from napistu.constants import (
     SOURCE_SPEC,
 )
 from napistu.identifiers import Identifiers
+from napistu.indices import PWIndex
 from napistu.ingestion import sbml
 from napistu.modify.cofactors import drop_cofactors
+from napistu.sbml_dfs_core import SBML_dfs
+from napistu.source import Source, create_source_table
 
 test_path = os.path.abspath(os.path.join(__file__, os.pardir))
 test_data = os.path.join(test_path, "test_data")
@@ -32,7 +46,7 @@ def test_reduce_to_consensus_ids():
     sbml_model = sbml.SBML(sbml_path)
     comp_species_df = sbml_model._define_cspecies()
     comp_species_df.index.names = [SBML_DFS.S_ID]
-    consensus_species, species_lookup = consensus._reduce_to_consensus_ids(
+    consensus_species, species_lookup = _reduce_to_consensus_ids(
         comp_species_df,
         {
             SCHEMA_DEFS.PK: SBML_DFS.S_ID,
@@ -48,10 +62,10 @@ def test_reduce_to_consensus_ids():
 
 
 def test_consensus():
-    pw_index = indices.PWIndex(os.path.join(test_data, SOURCE_SPEC.PW_INDEX_FILE))
-    sbml_dfs_dict = consensus.construct_sbml_dfs_dict(pw_index)
+    pw_index = PWIndex(os.path.join(test_data, SOURCE_SPEC.PW_INDEX_FILE))
+    sbml_dfs_dict = construct_sbml_dfs_dict(pw_index)
 
-    consensus_model = consensus.construct_consensus_model(sbml_dfs_dict, pw_index)
+    consensus_model = construct_consensus_model(sbml_dfs_dict, pw_index)
     assert consensus_model.species.shape == (38, 3)
     assert consensus_model.reactions.shape == (30, 4)
     assert consensus_model.reaction_species.shape == (137, 4)
@@ -78,7 +92,7 @@ def test_source_tracking():
             "new_id": [0, 0, 1, 1],
         }
     )
-    agg_tbl[table_schema[SCHEMA_DEFS.SOURCE]] = source.Source.empty()
+    agg_tbl[table_schema[SCHEMA_DEFS.SOURCE]] = Source.empty()
 
     # define new_ids and the models they came from
     # these models will be matched to the pw_index to flush out metadata
@@ -95,14 +109,14 @@ def test_source_tracking():
     )
 
     # use an existing pw_index since pw_index currently checks for the existence of the source file
-    pw_index = indices.PWIndex(os.path.join(test_data, "pw_index.tsv"))
+    pw_index = PWIndex(os.path.join(test_data, "pw_index.tsv"))
 
     # test create source table
-    source_table = source.create_source_table(lookup_table, table_schema, pw_index)
+    source_table = create_source_table(lookup_table, table_schema, pw_index)
     assert source_table["source_var"][0].source.shape == (2, 8)
 
     # test create_consensus_sources
-    consensus_sources = consensus._create_consensus_sources(
+    consensus_sources = _create_consensus_sources(
         agg_tbl, lookup_table, table_schema, pw_index
     )
     assert consensus_sources[0].source.shape == (2, 8)
@@ -122,7 +136,7 @@ def test_source_tracking():
 
     # expect a ValueError when the model is not found
     with pytest.raises(ValueError) as _:
-        source.create_source_table(invalid_lookup_table, table_schema, pw_index)
+        create_source_table(invalid_lookup_table, table_schema, pw_index)
 
     # now we will aggregate the consensus model above with a new single model (which has some
     # overlapping entries with the consensusd (id 1) and some new ids (id 2)
@@ -134,7 +148,7 @@ def test_source_tracking():
     )
 
     agg_tbl2[table_schema[SCHEMA_DEFS.SOURCE]] = consensus_sources.tolist() + [
-        source.Source.empty() for i in range(0, 2)
+        Source.empty() for i in range(0, 2)
     ]
 
     lookup_table2 = pd.DataFrame(
@@ -151,14 +165,14 @@ def test_source_tracking():
         }
     )
 
-    source_table = source.create_source_table(lookup_table2, table_schema, pw_index)
+    source_table = create_source_table(lookup_table2, table_schema, pw_index)
     assert source_table.shape == (3, 1)
     assert [
         source_table["source_var"][i].source.shape
         for i in range(0, source_table.shape[0])
     ] == [(1, 8), (2, 8), (1, 8)]
 
-    consensus_sources = consensus._create_consensus_sources(
+    consensus_sources = _create_consensus_sources(
         agg_tbl2, lookup_table2, table_schema, pw_index
     )
     assert [
@@ -168,8 +182,8 @@ def test_source_tracking():
 
 def test_passing_entity_data():
 
-    pw_index = indices.PWIndex(os.path.join(test_data, "pw_index.tsv"))
-    sbml_dfs_dict = consensus.construct_sbml_dfs_dict(pw_index)
+    pw_index = PWIndex(os.path.join(test_data, "pw_index.tsv"))
+    sbml_dfs_dict = construct_sbml_dfs_dict(pw_index)
 
     for model in list(sbml_dfs_dict.keys())[0:3]:
         sbml_dfs_dict[model].add_species_data(
@@ -191,7 +205,7 @@ def test_passing_entity_data():
 
     # create a consensus with perfect merges of overlapping id-table-variable values
     # i.e., when combined all merged entries have the same attributes
-    consensus_model = consensus.construct_consensus_model(sbml_dfs_dict, pw_index)
+    consensus_model = construct_consensus_model(sbml_dfs_dict, pw_index)
 
     assert len(consensus_model.species_data) == 1
     assert consensus_model.species_data["my_species_data"].shape == (10, 1)
@@ -208,7 +222,7 @@ def test_passing_entity_data():
             .to_frame(),
         )
 
-    consensus_model = consensus.construct_consensus_model(sbml_dfs_dict, pw_index)
+    consensus_model = construct_consensus_model(sbml_dfs_dict, pw_index)
     assert len(consensus_model.species_data) == 2
 
     # create a case where reactions will be merged and the same reaction
@@ -221,7 +235,7 @@ def test_passing_entity_data():
         0, SOURCE_SPEC.FILE
     ]
 
-    duplicated_sbml_dfs_dict = consensus.construct_sbml_dfs_dict(minimal_pw_index)
+    duplicated_sbml_dfs_dict = construct_sbml_dfs_dict(minimal_pw_index)
     # explicitely define the order we'll loop through models so that
     # the position of a model can be used to set mismatching attributes
     # for otherwise identical models
@@ -249,9 +263,7 @@ def test_passing_entity_data():
         "R-HSA-1237044"
     ].reactions.assign(r_isreversible=True)
 
-    consensus_model = consensus.construct_consensus_model(
-        duplicated_sbml_dfs_dict, pw_index
-    )
+    consensus_model = construct_consensus_model(duplicated_sbml_dfs_dict, pw_index)
     assert consensus_model.reactions_data["my_mismatched_data"].shape == (5, 3)
     assert consensus_model.reactions[SBML_DFS.R_ISREVERSIBLE].eq(True).all()
 
@@ -314,12 +326,8 @@ def test_report_consensus_merges_reactions(tmp_path, model_source_stub):
         SBML_DFS.REACTIONS: reactions,
         SBML_DFS.REACTION_SPECIES: reaction_species,
     }
-    sbml1 = sbml_dfs_core.SBML_dfs(
-        sbml_dict, model_source_stub, validate=False, resolve=False
-    )
-    sbml2 = sbml_dfs_core.SBML_dfs(
-        sbml_dict, model_source_stub, validate=False, resolve=False
-    )
+    sbml1 = SBML_dfs(sbml_dict, model_source_stub, validate=False, resolve=False)
+    sbml2 = SBML_dfs(sbml_dict, model_source_stub, validate=False, resolve=False)
     sbml_dfs_dict = {"mod1": sbml1, "mod2": sbml2}
 
     # Create a lookup_table that merges both reactions into a new_id
@@ -334,7 +342,7 @@ def test_report_consensus_merges_reactions(tmp_path, model_source_stub):
     table_schema = SBML_DFS_SCHEMA.SCHEMA[SBML_DFS.REACTIONS]
 
     # Call the function and check that it runs and the merge_labels are as expected
-    consensus._report_consensus_merges(
+    _report_consensus_merges(
         lookup_table.set_index([SOURCE_SPEC.MODEL, SBML_DFS.R_ID])[
             "new_id"
         ],  # this is a Series with name 'new_id'
@@ -380,8 +388,8 @@ def test_build_consensus_identifiers_handles_merges_and_missing_ids():
 
     schema = SBML_DFS_SCHEMA.SCHEMA[SBML_DFS.SPECIES]
 
-    indexed_cluster, cluster_consensus_identifiers = (
-        consensus._build_consensus_identifiers(df, schema)
+    indexed_cluster, cluster_consensus_identifiers = _build_consensus_identifiers(
+        df, schema
     )
 
     # All entities should be assigned to a cluster
@@ -431,10 +439,8 @@ def test_consensus_round_trip_consistency(
 
     # Path 2: sbml_dfs_dict -> sbml_dfs_list -> (sbml_dfs_dict, pw_index) -> consensus_model
     sbml_dfs_list = list(sbml_dfs_dict.values())
-    sbml_dfs_dict_from_list, pw_index_from_list = consensus.prepare_consensus_model(
-        sbml_dfs_list
-    )
-    sbml_dfs_from_list = consensus.construct_consensus_model(
+    sbml_dfs_dict_from_list, pw_index_from_list = prepare_consensus_model(sbml_dfs_list)
+    sbml_dfs_from_list = construct_consensus_model(
         sbml_dfs_dict_from_list, pw_index_from_list
     )
 
@@ -492,9 +498,7 @@ def test_pre_consensus_compartment_check_compatible(
 ):
     """Test that compatible models with overlapping compartments pass the check."""
     # This should not raise any errors or log warnings
-    consensus._pre_consensus_compartment_check(
-        sbml_dfs_dict_metabolism, pw_index_metabolism
-    )
+    _pre_consensus_compartment_check(sbml_dfs_dict_metabolism, pw_index_metabolism)
 
 
 def test_pre_consensus_compartment_check_incompatible(
@@ -519,11 +523,11 @@ def test_pre_consensus_compartment_check_incompatible(
     sbml_dfs_dict = {"fake_model": minimal_valid_sbml_dfs, "real_model": sbml_dfs}
 
     # Create pathway index object
-    pw_index = indices.PWIndex(pw_index_data, validate_paths=False)
+    pw_index = PWIndex(pw_index_data, validate_paths=False)
 
     # Capture log messages
     with caplog.at_level(logging.ERROR):
-        consensus._pre_consensus_compartment_check(sbml_dfs_dict, pw_index)
+        _pre_consensus_compartment_check(sbml_dfs_dict, pw_index)
 
     # Check that an error was logged about incompatible compartments
     assert len(caplog.records) > 0
@@ -533,7 +537,7 @@ def test_pre_consensus_compartment_check_incompatible(
 def test_pre_consensus_ontology_check_compatible(sbml_dfs_dict_metabolism):
     """Test that compatible models with overlapping ontologies pass the check."""
     # This should not raise any errors or log warnings
-    consensus._pre_consensus_ontology_check(sbml_dfs_dict_metabolism, SBML_DFS.SPECIES)
+    _pre_consensus_ontology_check(sbml_dfs_dict_metabolism, SBML_DFS.SPECIES)
 
 
 def test_pre_consensus_ontology_check_incompatible(
@@ -565,7 +569,7 @@ def test_pre_consensus_ontology_check_incompatible(
 
     # Capture log messages
     with caplog.at_level(logging.ERROR):
-        consensus._pre_consensus_ontology_check(sbml_dfs_dict, SBML_DFS.SPECIES)
+        _pre_consensus_ontology_check(sbml_dfs_dict, SBML_DFS.SPECIES)
 
     # Check that an error was logged about incompatible ontologies
     assert len(caplog.records) > 0
@@ -619,7 +623,7 @@ def test_update_foreign_keys():
     fk_lookup_tables = {SBML_DFS.S_ID: fk_lookup}
 
     # Test successful FK update
-    result = consensus._update_foreign_keys(agg_tbl, table_schema, fk_lookup_tables)
+    result = _update_foreign_keys(agg_tbl, table_schema, fk_lookup_tables)
 
     # Verify the foreign keys were updated correctly
     expected_species_ids = ["new_s1", "new_s2", "new_s1", "new_s3"]
@@ -656,7 +660,7 @@ def test_update_foreign_keys_missing_key():
         ValueError,
         match=f"keys from agg_tbl are missing from the {SBML_DFS.S_ID} lookup table",
     ):
-        consensus._update_foreign_keys(agg_tbl, table_schema, fk_lookup_tables)
+        _update_foreign_keys(agg_tbl, table_schema, fk_lookup_tables)
 
 
 def test_update_foreign_keys_multiple_fks():
@@ -700,7 +704,7 @@ def test_update_foreign_keys_multiple_fks():
     }
 
     # Test updating both FKs
-    result = consensus._update_foreign_keys(agg_tbl, table_schema, fk_lookup_tables)
+    result = _update_foreign_keys(agg_tbl, table_schema, fk_lookup_tables)
 
     # Verify both FKs were updated
     assert result[SBML_DFS.S_ID].tolist() == ["new_s1", "new_s2"]
@@ -716,7 +720,7 @@ def test_construct_consensus_model_no_rxn_pathway_ids(
     no_rxn_pathway_ids = ["tca"]  # Select TCA cycle as no-reaction pathway
 
     # Create consensus model with no_rxn_pathway_ids
-    consensus_model = consensus.construct_consensus_model(
+    consensus_model = construct_consensus_model(
         sbml_dfs_dict_metabolism,
         pw_index_metabolism,
         no_rxn_pathway_ids=no_rxn_pathway_ids,
@@ -735,7 +739,7 @@ def test_construct_consensus_model_no_rxn_pathway_ids(
 
     # Should raise ValueError when invalid pathway ID is provided
     with pytest.raises(ValueError) as exc_info:
-        consensus.construct_consensus_model(
+        construct_consensus_model(
             sbml_dfs_dict_metabolism,
             pw_index_metabolism,
             no_rxn_pathway_ids=invalid_no_rxn_pathway_ids,
